Fixed the z coord issues. Many things seem to have started working, including basic combat.
#givemoney and other ways of GAINING money now update the client and no longer require a zone. Losing money still needs a zone to update, for now.

+++ common/eq_packet_structs.h
@@ -1513,9 +1513,9 @@ struct TradeCoin_Struct{
 	uint32	amount;
 };
 struct TradeMoneyUpdate_Struct{
-	uint32	trader;
-	uint32	type;
-	uint32	amount;
+	uint16	trader;
+	uint16	type;
+	int32	amount;
 };
 /*
 ** Surname struct
